+++ sockeye/constants.py
@@ -25,6 +25,9 @@ UNK_SYMBOL = "<unk>"
 PAD_SYMBOL = "<pad>"
 PAD_ID = 0
 PAD_FORMAT = "<pad%d>"
+POINTER_FORMAT = "<ptr%d>"
+POINTER_PATTERN = "<ptr(?P<index>\d+)>"
+
 TOKEN_SEPARATOR = " "
 VOCAB_SYMBOLS = [PAD_SYMBOL, UNK_SYMBOL, BOS_SYMBOL, EOS_SYMBOL]
 # reserve extra space for the EOS or BOS symbol that is added to both source and target