+++ sockeye/decoder.py
@@ -97,7 +97,7 @@ class Decoder(ABC):
                         source_encoded_max_length: int,
                         target_embed: mx.sym.Symbol,
                         target_embed_lengths: mx.sym.Symbol,
-                        target_embed_max_length: int) -> mx.sym.Symbol:
+                        target_embed_max_length: int) -> Tuple[mx.sym.Symbol, Optional[mx.sym.Symbol]]:
         """
         Decodes a sequence of embedded target words and returns sequence of last decoder
         representations for each time step.
@@ -108,7 +108,9 @@ class Decoder(ABC):
         :param target_embed: Embedded target sequence. Shape: (batch_size, target_embed_max_length, target_num_embed).
         :param target_embed_lengths: Lengths of embedded target sequences. Shape: (batch_size,).
         :param target_embed_max_length: Dimension of the embedded target sequence.
-        :return: Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+        :return:
+            Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+            Pointer data. Shape: (batch_size, target_embed_max_length, source_seq_len).
         """
         pass
 
@@ -117,7 +119,7 @@ class Decoder(ABC):
                     step: int,
                     target_embed_prev: mx.sym.Symbol,
                     source_encoded_max_length: int,
-                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol]]:
+                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol], Optional[mx.sym.Symbol]]:
         """
         Decodes a single time step given the current step, the previous embedded target word,
         and previous decoder states.
@@ -128,7 +130,7 @@ class Decoder(ABC):
         :param target_embed_prev: Previous target word embedding. Shape: (batch_size, target_num_embed).
         :param source_encoded_max_length: Length of encoded source time dimension.
         :param states: Arbitrary list of decoder states.
-        :return: logit inputs, attention probabilities, next decoder states.
+        :return: logit inputs, attention probabilities, next decoder states, pointer scores.
         """
         pass
 
@@ -240,7 +242,7 @@ class TransformerDecoder(Decoder):
                         source_encoded_max_length: int,
                         target_embed: mx.sym.Symbol,
                         target_embed_lengths: mx.sym.Symbol,
-                        target_embed_max_length: int) -> mx.sym.Symbol:
+                        target_embed_max_length: int) -> Tuple[mx.sym.Symbol, Optional[mx.sym.Symbol]]:
         """
         Decodes a sequence of embedded target words and returns sequence of last decoder
         representations for each time step.
@@ -251,7 +253,9 @@ class TransformerDecoder(Decoder):
         :param target_embed: Embedded target sequence. Shape: (batch_size, target_embed_max_length, target_num_embed).
         :param target_embed_lengths: Lengths of embedded target sequences. Shape: (batch_size,).
         :param target_embed_max_length: Dimension of the embedded target sequence.
-        :return: Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+        :return:
+            Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+            Pointer data. Shape: (batch_size, target_embed_max_length, source_seq_len).
         """
 
         # (batch_size * heads, max_length)
@@ -273,13 +277,13 @@ class TransformerDecoder(Decoder):
             target = layer(target, target_bias, source_encoded, source_bias)
         target = self.final_process(target, None)
 
-        return target
+        return target, None
 
     def decode_step(self,
                     step: int,
                     target_embed_prev: mx.sym.Symbol,
                     source_encoded_max_length: int,
-                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol]]:
+                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol], Optional[mx.sym.Symbol]]:
         """
         Decodes a single time step given the current step, the previous embedded target word,
         and previous decoder states.
@@ -290,7 +294,7 @@ class TransformerDecoder(Decoder):
         :param target_embed_prev: Previous target word embedding. Shape: (batch_size, target_num_embed).
         :param source_encoded_max_length: Length of encoded source time dimension.
         :param states: Arbitrary list of decoder states.
-        :return: logit inputs, attention probabilities, next decoder states.
+        :return: logit inputs, attention probabilities, next decoder states, pointer scores.
         """
         # for step > 1, states contains source_encoded, source_encoded_lengths, and cache tensors.
         source_encoded, source_encoded_lengths, *cache = states  # type: ignore
@@ -329,7 +333,7 @@ class TransformerDecoder(Decoder):
         # TODO(fhieber): no attention probs for now
         attention_probs = mx.sym.sum(mx.sym.zeros_like(source_encoded), axis=2, keepdims=False)
 
-        return target, attention_probs, new_states
+        return target, attention_probs, new_states, None
 
     def _get_cache_per_layer(self, cache: List[mx.sym.Symbol]) -> List[Dict[str, Optional[mx.sym.Symbol]]]:
         """
@@ -557,7 +561,7 @@ class RecurrentDecoder(Decoder):
                         source_encoded_max_length: int,
                         target_embed: mx.sym.Symbol,
                         target_embed_lengths: mx.sym.Symbol,
-                        target_embed_max_length: int) -> mx.sym.Symbol:
+                        target_embed_max_length: int) -> Tuple[mx.sym.Symbol, Optional[mx.sym.Symbol]]:
         """
         Decodes a sequence of embedded target words and returns sequence of last decoder
         representations for each time step.
@@ -568,7 +572,9 @@ class RecurrentDecoder(Decoder):
         :param target_embed: Embedded target sequence. Shape: (batch_size, target_embed_max_length, target_num_embed).
         :param target_embed_lengths: Lengths of embedded target sequences. Shape: (batch_size,).
         :param target_embed_max_length: Dimension of the embedded target sequence.
-        :return: Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+        :return:
+            Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+            Pointer data. Shape: (batch_size, target_embed_max_length, source_seq_len).
         """
 
         # target_embed: target_seq_len * (batch_size, num_target_embed)
@@ -594,6 +600,8 @@ class RecurrentDecoder(Decoder):
 
         # hidden_all: target_embed_max_length * (batch_size, rnn_num_hidden)
         hidden_states = []  # type: List[mx.sym.Symbol]
+        attention_scores = []  # type: List[mx.sym.Symbol]
+        
         # TODO: possible alternative: feed back the context vector instead of the hidden (see lamtram)
         self.reset()
         for seq_idx in range(target_embed_max_length):
@@ -605,15 +613,18 @@ class RecurrentDecoder(Decoder):
                                                 seq_idx,
                                                 enc_last_hidden=enc_last_hidden)
             hidden_states.append(state.hidden)
+            attention_scores.append(attention_state.scores)
 
         # concatenate along time axis: (batch_size, target_embed_max_length, rnn_num_hidden)
-        return mx.sym.stack(*hidden_states, axis=1, name='%shidden_stack' % self.prefix)
+        hidden_stack = mx.sym.stack(*hidden_states, axis=1, name='%shidden_stack' % self.prefix)
+        attention_scores_stack = mx.sym.stack(*attention_scores, axis=1, name='%sattention_scores_stack' % self.prefix)
+        return hidden_stack, attention_scores_stack
 
     def decode_step(self,
                     step: int,
                     target_embed_prev: mx.sym.Symbol,
                     source_encoded_max_length: int,
-                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol]]:
+                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol], Optional[mx.sym.Symbol]]:
         """
         Decodes a single time step given the current step, the previous embedded target word,
         and previous decoder states.
@@ -624,7 +635,7 @@ class RecurrentDecoder(Decoder):
         :param target_embed_prev: Previous target word embedding. Shape: (batch_size, target_num_embed).
         :param source_encoded_max_length: Length of encoded source time dimension.
         :param states: Arbitrary list of decoder states.
-        :return: logit inputs, attention probabilities, next decoder states.
+        :return: logit inputs, attention probabilities, next decoder states, pointer scores.
         """
         source_encoded, prev_dynamic_source, source_encoded_length, prev_hidden, *layer_states = states
 
@@ -640,7 +651,7 @@ class RecurrentDecoder(Decoder):
 
         prev_state = RecurrentDecoderState(prev_hidden, list(layer_states))
         prev_attention_state = rnn_attention.AttentionState(context=None, probs=None,
-                                                            dynamic_source=prev_dynamic_source)
+                                                            dynamic_source=prev_dynamic_source, scores=None)
 
         # state.hidden: (batch_size, rnn_num_hidden)
         # attention_state.dynamic_source: (batch_size, source_seq_len, coverage_num_hidden)
@@ -656,7 +667,7 @@ class RecurrentDecoder(Decoder):
                       source_encoded_length,
                       state.hidden] + state.layer_states
 
-        return state.hidden, attention_state.probs, new_states
+        return state.hidden, attention_state.probs, new_states, attention_state.scores
 
     def reset(self):
         """
@@ -694,7 +705,7 @@ class RecurrentDecoder(Decoder):
         :return: List of symbolic initial states.
         """
         hidden, layer_states = self.get_initial_state(source_encoded, source_encoded_lengths)
-        context, attention_probs, dynamic_source = self.attention.get_initial_state(source_encoded_lengths,
+        context, attention_probs, dynamic_source, attention_scores = self.attention.get_initial_state(source_encoded_lengths,
                                                                                     source_encoded_max_length)
         states = [source_encoded, dynamic_source, source_encoded_lengths, hidden] + layer_states
         return states
@@ -1022,7 +1033,7 @@ class ConvolutionalDecoder(Decoder):
                         source_encoded_max_length: int,
                         target_embed: mx.sym.Symbol,
                         target_embed_lengths: mx.sym.Symbol,
-                        target_embed_max_length: int) -> mx.sym.Symbol:
+                        target_embed_max_length: int) -> Tuple[mx.sym.Symbol, Optional[mx.sym.Symbol]]:
         """
         Decodes a sequence of embedded target words and returns sequence of last decoder
         representations for each time step.
@@ -1033,7 +1044,9 @@ class ConvolutionalDecoder(Decoder):
         :param target_embed: Embedded target sequence. Shape: (batch_size, target_embed_max_length, target_num_embed).
         :param target_embed_lengths: Lengths of embedded target sequences. Shape: (batch_size,).
         :param target_embed_max_length: Dimension of the embedded target sequence.
-        :return: Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+        :return:
+            Decoder data. Shape: (batch_size, target_embed_max_length, decoder_depth).
+            Pointer data. Shape: (batch_size, target_embed_max_length, source_seq_len).
         """
 
         # (batch_size, target_seq_len, num_hidden)
@@ -1043,7 +1056,7 @@ class ConvolutionalDecoder(Decoder):
                                      target_embed_lengths=target_embed_lengths,
                                      target_embed_max_length=target_embed_max_length)
 
-        return target_hidden
+        return target_hidden, None
 
     def _decode(self,
                 source_encoded: mx.sym.Symbol,
@@ -1092,7 +1105,7 @@ class ConvolutionalDecoder(Decoder):
                     step: int,
                     target_embed_prev: mx.sym.Symbol,
                     source_encoded_max_length: int,
-                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol]]:
+                    *states: mx.sym.Symbol) -> Tuple[mx.sym.Symbol, mx.sym.Symbol, List[mx.sym.Symbol], Optional[mx.sym.Symbol]]:
         """
         Decodes a single time step given the current step, the previous embedded target word,
         and previous decoder states.
@@ -1103,7 +1116,7 @@ class ConvolutionalDecoder(Decoder):
         :param target_embed_prev: Previous target word embedding. Shape: (batch_size, target_num_embed).
         :param source_encoded_max_length: Length of encoded source time dimension.
         :param states: Arbitrary list of decoder states.
-        :return: logit inputs, attention probabilities, next decoder states.
+        :return: logit inputs, attention probabilities, next decoder states, pointer scores.
         """
         # Source_encoded: (batch_size, source_encoded_max_length, encoder_depth)
         source_encoded, source_encoded_lengths, *layer_states = states
@@ -1170,7 +1183,7 @@ class ConvolutionalDecoder(Decoder):
                                                            axis=2, begin=0, end=1),
                                          shape=(0, -1))
 
-        return target_hidden, attention_probs, [source_encoded, source_encoded_lengths] + new_layer_states
+        return target_hidden, attention_probs, [source_encoded, source_encoded_lengths] + new_layer_states, None
 
     def reset(self):
         pass