+++ test/unit/test_arguments.py
@@ -98,6 +98,7 @@ def test_device_args(test_params, expected_params):
               transformer_positional_embedding_type="fixed",
               transformer_preprocess=('n', 'n'),
               transformer_postprocess=('dr', 'dr'),
+              attention_based_copying=False,
               rnn_attention_use_prev_word=False,
               rnn_decoder_state_init="last",
               rnn_encoder_reverse_input=False,